script fix

+++ src/main/resources/db/migration/V1_000__create_cerveja.sql
@@ -3,7 +3,7 @@ CREATE TABLE TB_CERVEJA (
   marca varchar(200) NOT NULL,
   teor_alcoolico double DEFAULT NULL,
   tipo varchar(50) DEFAULT NULL,
-  preco double DEFAULT NULL,
+  preco decimal(19,2) DEFAULT NULL,
   data_lancamento timestamp DEFAULT NULL,
   created_date timestamp DEFAULT NULL,
   updated_date timestamp DEFAULT NULL,
